+++ src/Patient/Walkthrough/Outro.js
