+++ src/Patient/Walkthrough/PaginationDot.js
